Add feature User Info
