+++ Workhour/Core/Profile/ProfileView.swift
@@ -7,55 +7,55 @@
 
 import SwiftUI
 
-@MainActor
-final class ProfileViewModel: ObservableObject{
-    
-    @Published private(set) var user : DBUser? = nil
-    
-    func loadCurrnetUser() async throws {
-        guard let authDataResult = try AuthenticationManager.shared.getAuthenticatedUser() else { return }
-        self.user = try await UserManager.shared.getUser(userId: authDataResult.uid)
-    }
-}
-
-
 struct ProfileView: View {
     @StateObject private var viewModel = ProfileViewModel()
     @Binding var showSignInView : Bool
     
     var body: some View {
-        List{
-            if let user = viewModel.user{
-                Text("UserId: \(user.userId)")
-                
-                if let email = user.email{
-                    Text("email: \(email.description)")
+        VStack{
+            
+            //USER INFO
+            HStack(spacing: 20){
+                if let user = viewModel.user{
+                    if let photoUrl = user.photoUrl{
+                        URLImageView(urlString: photoUrl)
+                        
+                        
+                        VStack(alignment: .leading){
+                            Text("Good Morning")
+                            if let name = user.name{
+                                Text("\(name)")
+                            }
+                        }
+
+                    }
                 }
-                
-                if let name = user.name{
-                    Text("name: \(name)")
-                }
-                
-                if let photoUrl = user.photoUrl{
-                    URLImageView(urlString: photoUrl)
-                }
-                
+                Spacer()
             }
+
+            
+            //
+            
+            
+            
         }
         .task{
             try? await viewModel.loadCurrnetUser()
         }
+
+
+
         .navigationTitle("Profile")
-        .toolbar{
-            ToolbarItem(placement: .navigationBarTrailing) {
-                NavigationLink {
-                    SettingView(showSignInView: $showSignInView)
-                } label: {
-                    Image(systemName: "gear")
-                        .font(.headline)
-                }
-            }
-        }
+        //        .toolbar{
+        //            ToolbarItem(placement: .navigationBarTrailing) {
+        //                NavigationLink {
+        //                    SettingView(showSignInView: $showSignInView)
+        //                } label: {
+        //                    Image(systemName: "gear")
+        //                        .font(.headline)
+        //                }
+        //            }
+        //        }
     }
 }
 
@@ -71,19 +71,19 @@ import Combine
 class ImageLoader: ObservableObject {
     @Published var image: UIImage?
     private var cancellable: AnyCancellable?
-
+    
     func load(fromURLString urlString: String) {
         guard let url = URL(string: urlString) else {
             return
         }
-
+        
         cancellable = URLSession.shared.dataTaskPublisher(for: url)
             .map { UIImage(data: $0.data) }
             .replaceError(with: nil)
             .receive(on: DispatchQueue.main)
             .sink { [weak self] in self?.image = $0 }
     }
-
+    
     deinit {
         cancellable?.cancel()
     }
@@ -92,13 +92,15 @@ class ImageLoader: ObservableObject {
 struct URLImageView: View {
     @StateObject private var imageLoader = ImageLoader()
     let urlString: String
-
+    
     var body: some View {
         Group {
             if let image = imageLoader.image {
                 Image(uiImage: image)
                     .resizable()
-                    .aspectRatio(contentMode: .fit)
+                    .aspectRatio(contentMode: .fill)
+                    .frame(width: 50, height: 50)
+                    .clipShape(Circle())
             } else {
                 Text("Loading...")
             }
